refactor: remove useless service
